renamed internal struct to follow naming convention + some compile fixes on newer compilers

+++ src/dl_reflect.cpp
@@ -103,7 +103,7 @@ dl_error_t DL_DLL_EXPORT dl_reflect_get_type_members( dl_ctx_t dl_ctx, dl_typeid
 					out_members[nMember].array_count = Member.size[DL_PTR_SIZE_HOST] / pSubType->size[DL_PTR_SIZE_HOST];
 				}
 				break;
-				case DL_TYPE_STORAGE_STR: out_members[nMember].array_count = Member.size[DL_PTR_SIZE_HOST] / sizeof(char*); break;
+				case DL_TYPE_STORAGE_STR: out_members[nMember].array_count = (unsigned int)(Member.size[DL_PTR_SIZE_HOST] / sizeof(char*)); break;
 				default:
 					out_members[nMember].array_count = Member.size[DL_PTR_SIZE_HOST] / (uint32_t)DLPodSize(Member.type); break;
 			}
